pyproject.toml
[project]
authors = [{ name = "nypox", email = "hello@nypox.com" }]
name = "quite"
requires-python = "== 3.10"
version = "0.1.0"
dependencies = [
    "numpy==1.26.4",
    "loguru>=0.7.2",
    "omegaconf>=2.3.0",
    "msgspec==0.19.0",
    "requests>=2.31.0",
    "Pillow>=10.2.0",
    "pillow-avif-plugin>=1.4.2",
    "nanoid>=2.0.0",
    "asyncio==3.4.3",
    "aiohttp==3.11.2",
    "grpcio==1.67.1",
    "grpcio-reflection==1.67.1",
    "opencv-python-headless>=4.9.0.80",
    "soundfile>=0.12.1",
    "librosa>=0.10.2.post1",
    "setuptools>=80.4.0",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
quite = { path = ".", editable = true }

[tool.pixi.tasks]
